requirements.txt
crewai
langchain
openai
python-dotenv
pypdf2
duckduckgo-search
sqlalchemy
streamlit
gitpython
playwright
beautifulsoup4
requests
TTS
soundfile
sounddevice
numpy
torch
torchaudio